Build client differently; clean up makefile
I had previously been building the wasm client the recommended way,
but it seemed to be having trouble once I tried also compiling a
server with the same Cargo file. Most likely I "should" be using
separate Cargo files and maybe workspaces, but I instead changed the
build to follow https://github.com/rust-lang/cargo/issues/4881 , which
says to build the library as "rlib" instead of "cdylib", and then
build a wasm binary out of it.

I also debugged and fixed the Makefile rules:

* Changed the hard-coded path to my own directory structure to instead
  always build the wasm to build/. I can make this a symlink if I want
  to point to my own directory structure instead.
* Changed the servers to use the debug build, but with -O1. For this
  project I don't need a release build.

I added a readme file in preparation for uploading to github.