Move JS websocket code into a class
I considered making the class available on the Rust side, as described
on [1], so that the Rust code could open the connection and work with
it. However, that complicates the callbacks from the JS side to the
Rust side, and it's just not that important for this demo, so I made
the connection a global.

[1]: https://rustwasm.github.io/docs/wasm-bindgen/examples/import-js.html